backend/requirements.txt
uvicorn==0.27.1
python-dotenv==1.0.1
Werkzeug==2.2.2
marshmallow==3.14.1
pymongo==4.0.1
fastapi==0.110.0
bcrypt==4.1.2
pydantic>=2.6.3
PyJWT==2.8.0